Add server version command
Add a server version command, that displays the version of the remote debugger.

That is usefull, because it provides an easy method to check if a potential problem is
caused by a too old remote debugger.

Now the cofigure script on linux and a prebuild-step in MS Visual Studio generate a file
named src/version.h containing a macro called JSRDBG_VERSION. This macro is defined by
the output of 'git describe --tags --always --dirty'